more ints to floats

+++ EliteLib/Structs/EquipmentItem.cs
@@ -7,11 +7,11 @@
         internal bool CanBuy;
         internal bool Show;
         internal int TechLevel;
-        internal int Price;
+        internal float Price;
         internal string Name;
         internal EquipmentType Type;
 
-        internal EquipmentItem(bool canBuy, bool show, int techLevel, int price, string name, EquipmentType type)
+        internal EquipmentItem(bool canBuy, bool show, int techLevel, float price, string name, EquipmentType type)
         {
             CanBuy = canBuy;
             Show = show;
