+++ EliteLib/Structs/planet_data.cs
@@ -5,7 +5,7 @@
         internal int government;
         internal int economy;
         internal int techlevel;
-        internal int population;
+        internal float population;
         internal int productivity;
         internal int radius;
     };
